Refactor user to different files
