+++ pkg/framework/objects/user/data_source.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
@@ -23,37 +22,12 @@ type userDataSource struct {
 	client *dbt_cloud.Client
 }
 
-type userDataSourceModel struct {
-	ID    types.Int64  `tfsdk:"id"`
-	Email types.String `tfsdk:"email"`
-}
-
 func (d *userDataSource) Metadata(
 	_ context.Context,
 	req datasource.MetadataRequest,
 	resp *datasource.MetadataResponse,
 ) {
 	resp.TypeName = req.ProviderTypeName + "_user"
-}
-
-func (d *userDataSource) Schema(
-	_ context.Context,
-	_ datasource.SchemaRequest,
-	resp *datasource.SchemaResponse,
-) {
-	resp.Schema = schema.Schema{
-		Description: "Retrieve user details",
-		Attributes: map[string]schema.Attribute{
-			"id": schema.Int64Attribute{
-				Computed:    true,
-				Description: "ID of the user",
-			},
-			"email": schema.StringAttribute{
-				Required:    true,
-				Description: "Email for the user",
-			},
-		},
-	}
 }
 
 func (d *userDataSource) Read(
